optimize cache trait

+++ src/traits/CacheTrait.php
@@ -9,30 +9,35 @@ use tpr\Files;
 
 trait CacheTrait
 {
+    private static array $cache_data = [];
+
     private function cache(string $tmp_file, ?array $tmp_data = null): ?array
     {
-        // count cache file name
+        if (true === App::debugMode()) {
+            return null;
+        }
         $cache_time = App::drive()->getConfig()->cache_time;
-        $count      = (int) (time() / $cache_time);
+        $count      = 0 === $cache_time ? 'cache' : (int) (time() / $cache_time);
         $cache_file = $tmp_file . \DIRECTORY_SEPARATOR . (string) $count . '.php';
         if (null === $tmp_data) {
-            if (true === App::debugMode()) {
-                return null;
-            }
-            if (!file_exists($cache_file)) {
-                Files::remove($tmp_file);
+            if (!isset(self::$cache_data[$tmp_file])) {
+                if (!file_exists($cache_file)) {
+                    Files::remove($tmp_file);
 
-                return null;
+                    return null;
+                }
+                self::$cache_data[$tmp_file] = require_once $cache_file;
+
+                return self::$cache_data[$tmp_file];
             }
 
-            return require_once $cache_file;
+            return self::$cache_data[$tmp_file];
         }
-
-        if (!App::debugMode()) {
+        if (!isset(self::$cache_data[$tmp_file])) {
+            self::$cache_data[$tmp_file] = $tmp_data;
             Files::save($cache_file, '<?php' . PHP_EOL . 'return ' . var_export($tmp_data, true) . ';' . PHP_EOL);
-        } else {
-            Files::remove($cache_file);
         }
+
         unset($cache_file, $cache_time, $count);
 
         return $tmp_data;
